# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Study Group Finder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  faculty: string;
  academicYear: number;
  subjectsOfInterest?: string[];
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  faculty: string;
  academicYear: number;
  subjectsOfInterest?: string[];
  isBanned: boolean;
  createdAt?: string;
}

export type UserWithRoleRole =
  (typeof UserWithRoleRole)[keyof typeof UserWithRoleRole];

export const UserWithRoleRole = {
  student: "student",
  admin: "admin",
} as const;

export type UserWithRole = User & {
  role: UserWithRoleRole;
};

export interface AuthResponse {
  user: UserWithRole;
  token: string;
}

export interface UpdateProfileBody {
  name?: string;
  faculty?: string;
  academicYear?: number;
  subjectsOfInterest?: string[];
}

export type GroupType = (typeof GroupType)[keyof typeof GroupType];

export const GroupType = {
  online: "online",
  offline: "offline",
} as const;

export interface Group {
  id: number;
  title: string;
  subject: string;
  description: string;
  type: GroupType;
  location?: string | null;
  dateTime: string;
  maxMembers: number;
  createdBy: number;
  createdAt?: string;
}

export type GroupWithMemberCount = Group & {
  memberCount: number;
  creator?: User;
  isMember?: boolean;
};

export type GroupDetail = GroupWithMemberCount & {
  members: User[];
};

export type CreateGroupBodyType =
  (typeof CreateGroupBodyType)[keyof typeof CreateGroupBodyType];

export const CreateGroupBodyType = {
  online: "online",
  offline: "offline",
} as const;

export interface CreateGroupBody {
  title: string;
  subject: string;
  description: string;
  type: CreateGroupBodyType;
  location?: string | null;
  dateTime: string;
  maxMembers: number;
}

export interface Message {
  id: number;
  groupId: number;
  userId: number;
  content: string;
  isFlagged: boolean;
  createdAt: string;
  user?: User;
}

export type AdminMessage = Message & {
  groupTitle: string;
};

export interface SendMessageBody {
  content: string;
}

export type ReportType = (typeof ReportType)[keyof typeof ReportType];

export const ReportType = {
  user: "user",
  group: "group",
  message: "message",
} as const;

export type ReportStatus = (typeof ReportStatus)[keyof typeof ReportStatus];

export const ReportStatus = {
  pending: "pending",
  resolved: "resolved",
} as const;

export interface Report {
  id: number;
  reporterId: number;
  type: ReportType;
  targetId: number;
  reason: string;
  status: ReportStatus;
  createdAt: string;
}

export type ReportWithReporter = Report & {
  reporter?: User;
};

export type CreateReportBodyType =
  (typeof CreateReportBodyType)[keyof typeof CreateReportBodyType];

export const CreateReportBodyType = {
  user: "user",
  group: "group",
  message: "message",
} as const;

export interface CreateReportBody {
  type: CreateReportBodyType;
  targetId: number;
  reason: string;
}

export interface BanUserBody {
  banned: boolean;
}

export type ChangeRoleBodyRole =
  (typeof ChangeRoleBodyRole)[keyof typeof ChangeRoleBodyRole];

export const ChangeRoleBodyRole = {
  student: "student",
  admin: "admin",
} as const;

export interface ChangeRoleBody {
  role: ChangeRoleBodyRole;
}

export interface AdminLogEntry {
  id: number;
  adminId: number;
  action: string;
  targetType: string;
  targetId: number;
  createdAt: string;
  admin?: User;
}

export type DashboardSummarySubjectBreakdownItem = {
  subject: string;
  count: number;
};

export interface DashboardSummary {
  totalGroups: number;
  totalMembers: number;
  onlineGroups: number;
  offlineGroups: number;
  myGroupsCount: number;
  upcomingCount: number;
  subjectBreakdown: DashboardSummarySubjectBreakdownItem[];
}

export interface AdminStats {
  totalUsers: number;
  totalGroups: number;
  totalMessages: number;
  activeGroupsToday: number;
  newUsersThisWeek: number;
  newGroupsThisWeek: number;
  totalReports: number;
  pendingReports: number;
  bannedUsers: number;
}

export type GetGroupsParams = {
  subject?: string;
  type?: GetGroupsType;
  date?: string;
  search?: string;
};

export type GetGroupsType = (typeof GetGroupsType)[keyof typeof GetGroupsType];

export const GetGroupsType = {
  online: "online",
  offline: "offline",
} as const;

export type GetMessagesParams = {
  limit?: number;
  before?: number;
};

export type AdminGetMessagesParams = {
  groupId?: number;
  flagged?: boolean;
  limit?: number;
  offset?: number;
};

export type AdminGetReportsParams = {
  status?: AdminGetReportsStatus;
};

export type AdminGetReportsStatus =
  (typeof AdminGetReportsStatus)[keyof typeof AdminGetReportsStatus];

export const AdminGetReportsStatus = {
  pending: "pending",
  resolved: "resolved",
} as const;

export type AdminGetLogsParams = {
  limit?: number;
  offset?: number;
};