# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Study Group Finder API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users/{id}:
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Delete a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/users/{id}/ban:
    patch:
      operationId: adminBanUser
      tags: [admin]
      summary: Ban or unban a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanUserBody"
      responses:
        "200":
          description: Ban status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/users/{id}/role:
    patch:
      operationId: adminChangeRole
      tags: [admin]
      summary: Change a user's role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeRoleBody"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/groups/{id}:
    delete:
      operationId: adminDeleteGroup
      tags: [admin]
      summary: Delete a group (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/messages:
    get:
      operationId: adminGetMessages
      tags: [admin]
      summary: Get all messages with pagination (admin only)
      parameters:
        - name: groupId
          in: query
          schema:
            type: integer
        - name: flagged
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: All messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminMessage"
  /admin/messages/{id}:
    delete:
      operationId: adminDeleteMessage
      tags: [admin]
      summary: Delete a message (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Message deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/reports:
    get:
      operationId: adminGetReports
      tags: [admin]
      summary: Get all reports (admin only)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, resolved]
      responses:
        "200":
          description: All reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReportWithReporter"
  /admin/reports/{id}/resolve:
    patch:
      operationId: adminResolveReport
      tags: [admin]
      summary: Resolve a report (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/logs:
    get:
      operationId: adminGetLogs
      tags: [admin]
      summary: Get admin audit logs (admin only)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminLogEntry"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        faculty:
          type: string
        academicYear:
          type: integer
        subjectsOfInterest:
          type: array
          items:
            type: string
        isBanned:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - faculty
        - academicYear
        - isBanned
...
    Message:
      type: object
      properties:
        id:
          type: integer
        groupId:
          type: integer
        userId:
          type: integer
        content:
          type: string
        isFlagged:
          type: boolean
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
      required:
        - id
        - groupId
        - userId
        - content
        - isFlagged
        - createdAt
    AdminMessage:
      allOf:
        - $ref: "#/components/schemas/Message"
        - type: object
          properties:
            groupTitle:
              type: string
          required:
            - groupTitle
...
    Report:
      type: object
      properties:
        id:
          type: integer
        reporterId:
          type: integer
        type:
          type: string
          enum: [user, group, message]
        targetId:
          type: integer
        reason:
          type: string
        status:
          type: string
          enum: [pending, resolved]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - reporterId
        - type
        - targetId
        - reason
        - status
        - createdAt
    ReportWithReporter:
      allOf:
        - $ref: "#/components/schemas/Report"
        - type: object
          properties:
            reporter:
              $ref: "#/components/schemas/User"
...
    BanUserBody:
      type: object
      properties:
        banned:
          type: boolean
      required:
        - banned
    ChangeRoleBody:
      type: object
      properties:
        role:
          type: string
          enum: [student, admin]
      required:
        - role
    AdminLogEntry:
      type: object
      properties:
        id:
          type: integer
        adminId:
          type: integer
        action:
          type: string
        targetType:
          type: string
        targetId:
          type: integer
        createdAt:
          type: string
          format: date-time
        admin:
          $ref: "#/components/schemas/User"
      required:
        - id
        - adminId
        - action
        - targetType
        - targetId
        - createdAt
...