# lib/api-zod/src/generated/types/adminMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Study Group Finder API
 * OpenAPI spec version: 0.1.0
 */
import type { Message } from "./message";

export type AdminMessage = Message & {
  groupTitle: string;
};