# lib/api-zod/src/generated/types/changeRoleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Study Group Finder API
 * OpenAPI spec version: 0.1.0
 */
import type { ChangeRoleBodyRole } from "./changeRoleBodyRole";

export interface ChangeRoleBody {
  role: ChangeRoleBodyRole;
}