# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Study Group Finder API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  email: string;
  faculty: string;
  academicYear: number;
  subjectsOfInterest?: string[];
  isBanned: boolean;
  createdAt?: Date;
}